delete field from part wip
