+++ DFC.ServiceTaxonomy.GraphSync/GraphSyncers/Interfaces/IGraphResyncer.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace DFC.ServiceTaxonomy.GraphSync.GraphSyncers.Interfaces
+{
+    public interface IGraphResyncer
+    {
+        Task ResyncContentItems(string contentType);
+    }
+}
